backend/requirement.txt
fastapi
uvicorn
pydantic
openai
requests
python-dotenv
pytest
httpx